Module google_clouduseraccountsvm_beta::api

Structs

AuthorizedKeysView
A list of authorized public keys for a user account.

CloudUserAccounts
Central instance to access all CloudUserAccounts related resource activities

GlobalAccountsOperationDeleteCall
Deletes the specified operation resource.

GlobalAccountsOperationGetCall
Retrieves the specified operation resource.

GlobalAccountsOperationListCall
Retrieves the list of operation resources contained within the specified project.

GlobalAccountsOperationMethods
A builder providing access to all methods supported on globalAccountsOperation resources. It is not used directly, but through the CloudUserAccounts hub.

Group
A Group resource.

GroupAddMemberCall
Adds users to the specified group.

GroupDeleteCall
Deletes the specified Group resource.

GroupGetCall
Returns the specified Group resource.

GroupInsertCall
Creates a Group resource in the specified project using the data included in the request.

GroupList
There is no detailed description.

GroupListCall
Retrieves the list of groups contained within the specified project.

GroupMethods
A builder providing access to all methods supported on group resources. It is not used directly, but through the CloudUserAccounts hub.

GroupRemoveMemberCall
Removes users from the specified group.

GroupsAddMemberRequest
There is no detailed description.

GroupsRemoveMemberRequest
There is no detailed description.

LinuxAccountViews
A list of all Linux accounts for this project. This API is only used by Compute Engine virtual machines to get information about user accounts for a project or instance. Linux resources are read-only views into users and groups managed by the Compute Engine Accounts API.

LinuxGetAuthorizedKeysViewCall
Returns a list of authorized public keys for a specific user account.

LinuxGetAuthorizedKeysViewResponse
There is no detailed description.

LinuxGetLinuxAccountViewCall
Retrieves a list of user accounts for an instance within a specific project.

LinuxGetLinuxAccountViewsResponse
There is no detailed description.

LinuxGroupView
A detailed view of a Linux group.

LinuxMethods
A builder providing access to all methods supported on linux resources. It is not used directly, but through the CloudUserAccounts hub.

LinuxUserView
A detailed view of a Linux user account.

Operation
An Operation resource, used to manage asynchronous API requests.

OperationError
[Output Only] If errors are generated during processing of the operation, this field will be populated.

OperationErrorErrors
[Output Only] The array of errors encountered while processing this operation.

OperationList
Contains a list of Operation resources.

OperationWarnings
[Output Only] If warning messages are generated during processing of the operation, this field will be populated.

OperationWarningsData
[Output Only] Metadata about this warning in key: value format. For example: "data": [ { "key": "scope", "value": "zones/us-east1-d" }

PublicKey
A public key for authenticating to guests.

User
A User resource.

UserAddPublicKeyCall
Adds a public key to the specified User resource with the data included in the request.

UserDeleteCall
Deletes the specified User resource.

UserGetCall
Returns the specified User resource.

UserInsertCall
Creates a User resource in the specified project using the data included in the request.

UserList
There is no detailed description.

UserListCall
Retrieves a list of users contained within the specified project.

UserMethods
A builder providing access to all methods supported on user resources. It is not used directly, but through the CloudUserAccounts hub.

UserRemovePublicKeyCall
Removes the specified public key from the user.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.