Struct google_clouduseraccountsvm_beta::Group

pub struct Group {
    pub kind: Option<String>,
    pub name: Option<String>,
    pub members: Option<Vec<String>>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub description: Option<String>,
}

A Group resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete groups (none)

remove member groups (none)

get groups (response)

add member groups (none)

insert groups (request)

list groups (none)

Fields
kind: Option<String>

[Output Only] Type of the resource. Always clouduseraccounts#group for groups.

name: Option<String>

Name of the resource; provided by the client when the resource is created.

members: Option<Vec<String>>

[Output Only] A list of URLs to User resources who belong to the group. Users may only be members of groups in the same project.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

id: Option<String>

[Output Only] Unique identifier for the resource; defined by the server.

self_link: Option<String>

[Output Only] Server defined URL for the resource.

description: Option<String>

An optional textual description of the resource; provided by the client when the resource is created.