Struct google_clouduseraccountsvm_beta::UserMethods

pub struct UserMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on user resources. It is not used directly, but through the CloudUserAccounts hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_clouduseraccountsvm_beta as clouduseraccountsvm_beta;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use clouduseraccountsvm_beta::CloudUserAccounts;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CloudUserAccounts::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `add_public_key(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)` and `remove_public_key(...)`
// to build up your call.
let rb = hub.users();

Methods

impl<'a, C, A> UserMethods<'a, C, A>

fn add_public_key(&self, request: PublicKey, project: &str, user: &str) -> UserAddPublicKeyCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a public key to the specified User resource with the data included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

user - Name of the user for this request.

fn list(&self, project: &str) -> UserListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of users contained within the specified project.

Arguments

project - Project ID for this request.

fn insert(&self, request: User, project: &str) -> UserInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a User resource in the specified project using the data included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

fn get(&self, project: &str, user: &str) -> UserGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the specified User resource.

Arguments

project - Project ID for this request.

user - Name of the user resource to return.

fn remove_public_key(&self, project: &str, user: &str, fingerprint: &str) -> UserRemovePublicKeyCall<'a, C, A>

Create a builder to help you perform the following task:

Removes the specified public key from the user.

Arguments

project - Project ID for this request.

user - Name of the user for this request.

fingerprint - The fingerprint of the public key to delete. Public keys are identified by their fingerprint, which is defined by RFC4716 to be the MD5 digest of the public key.

fn delete(&self, project: &str, user: &str) -> UserDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified User resource.

Arguments

project - Project ID for this request.

user - Name of the user resource to delete.